% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{KSdatagen}
\alias{KSdatagen}
\title{Generate a dataset from Kang & Schafer simulation}
\usage{
KSdatagen(n)
}
\arguments{
\item{n}{Sample size}
}
\value{
A data.frame with n rows
}
\description{
Generate a dataset from Kang & Schafer simulation
}
